requirements.txt
# Add your project dependencies here
seaborn
matplotlib
pandas
tqdm 
sentence-transformers 
faiss-cpu 
langchain 
numpy
gradio
"transformers[torch]"